+++ tests/kubernetes/checks/example_ApiServerRequestTimeout/api-server-secure-port-PASSED.yaml
@@ -11,7 +11,7 @@ spec:
   containers:
   - command:
     - kube-apiserver
-    - --secure-port=80
+    - --request-timeout=2m3s
     image: gcr.io/google_containers/kube-apiserver-amd64:v1.6.0
     livenessProbe:
       failureThreshold: 8
@@ -23,3 +23,25 @@ spec:
       initialDelaySeconds: 15
       timeoutSeconds: 15
     name: kube-apiserver
+    resources:
+      requests:
+        cpu: 250m
+    volumeMounts:
+    - mountPath: /etc/kubernetes/
+      name: k8s
+      readOnly: true
+    - mountPath: /etc/ssl/certs
+      name: certs
+    - mountPath: /etc/pki
+      name: pki
+  hostNetwork: true
+  volumes:
+  - hostPath:
+      path: /etc/kubernetes
+    name: k8s
+  - hostPath:
+      path: /etc/ssl/certs
+    name: certs
+  - hostPath:
+      path: /etc/pki
+    name: pki
